demo/distributed_crawler/requirements.txt
scrapy==2.5.1
scrapy-redis==0.7.1
redis==4.0.2
pymongo==4.0.1
elasticsearch==7.17.0
bloom-filter==1.3
dnspython==2.2.1
prometheus-client==0.14.1 